turn off xcb request-check for CompositeGlyphs32
Included is a TODO to allow turning it on if debugging is enabled.

This request currently happens too often since the glyph stream is not yet
growable. In any case it happens often enough that there’s no reason to do the
request check unless it needs to be observed for testing.