; version 0.21.3
Update fontsloth dependecy to 0.15.1.

Add a date module:

By default it displays a calendar icon; abbreviated date as weekday, month,
day; and a local 12 hour time. The time format specifiers are the same as for
Emacs format-time-string. The icon and date change colors with the
seasons. Seasonal colors are customizable.

Bug fix:
+ check for uninitialized modules in watch-modules

  This should ensure that the module list can be safely updated with a newly
  created module while exlybar is running.